+++ dev/src/clasew/examples4.clj
@@ -13,9 +13,8 @@
 
 ;;; Demonstrate using Excel as a data store
 
-;; ---------------- RAW CALLING -----------------------------------------------
+;; ---------------- BASIC CALLING ---------------------------------------------
 
-;; Demo basic calling
 
 (def sample1 (es/clasew-excel-script "clasew-ex4-sample1.xlsx"
       es/create es/no-open "path to desktop"
@@ -24,14 +23,14 @@
 ; (p sample1)
 ; (p (es/clasew-excel-call! sample1))
 
-;; Demo Handler Chaining - Create, put values, get first row, save, quit
+;; Handler Chaining - Create, put values, get first row, save, quit
 
 (def datum (into []
       (map vec
            (for [i (range 10)]
              (repeatedly 10 #(rand-int 100))))))
 
-(p datum)
+;(p datum)
 
 (def sample2 (es/clasew-excel-script "clasew-ex4-sample2.xlsx"
       es/create es/no-open "path to desktop"
@@ -39,8 +38,8 @@
                              [:get-range-data "Sheet1" "A1:A10"]
                              [:save-quit]])))
 
-(p sample2)
-(p (es/clasew-excel-call! sample2))
+;(p sample2)
+;(p (es/clasew-excel-call! sample2))
 
 (def sample2a (es/clasew-excel-script "clasew-ex4-sample2.xlsx"
       es/no-create es/open "path to desktop"
@@ -48,8 +47,8 @@
                                 [:quit-no-save]])))
 
 
-(p sample2a)
-(p (es/clasew-excel-call! sample2a))
+;(p sample2a)
+;(p (es/clasew-excel-call! sample2a))
 ;; Demo - Open, Read, Quit
 
 (def sample3 (es/clasew-excel-script "clasew-ex4-sample2.xlsx"
@@ -85,9 +84,61 @@
                      [:save-quit]))
 
 
-;(def s5r (es/clasew-excel-call! sample5))
+; (def s5r (es/clasew-excel-call! sample5))
 
-;(p (es/clean-excel-result s5r))
+; (p (es/chain-put-range-data "Sheet1" datum))
+
+;; Demo misc forms
+
+
+; Uniform collection
+(def uniform [[1 2 3] [4 5 6] [7 8 9]])
+
+; Jagged collection
+(def jagged [[1] [2 2] [3 3 3]])
+
+(es/get-excel-range-a1 [0 0 0 9])
+(es/get-excel-range-a1 [0 1 1 9])
+
+(es/formula-wrap "SUM" es/get-excel-range-a1 [0 0 0 9])
+(es/formula-wrap "AVG" es/get-excel-range-a1 [0 0 0 9])
+
+(es/fsum es/get-excel-range-a1 [0 0 0 9])
+(es/favg es/get-excel-range-a1 [0 0 0 9])
+
+(es/row-ranges uniform)
+(es/row-ranges jagged)
+(es/row-ranges (es/pad-rows jagged))
+
+(es/column-ranges uniform)
+(es/column-ranges jagged)
+(es/column-ranges (es/pad-rows jagged))
+
+(map es/get-excel-range-a1 (es/column-ranges uniform))
+(map es/get-excel-range-a1 (es/row-ranges uniform 4 3))
+(map es/get-excel-range-a1 (es/row-ranges jagged))
+
+(es/sum-by-row uniform)
+(es/sum-by-col uniform)
+
+(es/avg-by-row uniform)
+(es/avg-by-col uniform)
+
+(es/extend-rows uniform 0 0 es/sum-by-row)
+(es/extend-columns uniform 0 0 es/avg-by-col)
+
+;; Put it together
+
+(def sample5a (es/create-wkbk "clasew-ex4-sample5a.xlsx" wrkbk-path
+      (es/chain-put-range-data "Sheet1"
+                               (es/extend-columns
+                                (es/extend-rows
+                                 datum 0 0 es/sum-by-row es/avg-by-row)
+                                0 0 es/avg-by-col))
+                               [:save-quit]))
+
+#_(p sample5a)
+#_(p (es/clasew-excel-call! sample5a))
 
 ;; Demo - cleaning up results