+++ src/clasew/excel.clj
@@ -47,7 +47,10 @@
    :run-script          "clasew_excel_run"
    })
 
+
+;;
 ;; Pure helpers
+;;
 
 (defn modify-keys
   "Uses zipmap to process 'f' on keys
@@ -74,23 +77,60 @@
   [{:keys [result] :as return-map}]
   (assoc return-map :result (w/prewalk ocm result)))
 
-(def ^:const ^:private base-char 65)
+(def ^:const ^:private base-divisor 26)          ; letters in ascii alphabet
+(def ^:const ^:private base-char 65)             ; "A"
+(def ^:const ^:private excel-max-cols 16384)     ; OS X MS 2011
+(def ^:const ^:private excel-max-rows 1048576)   ; OS X MS 2011
 
 (defn get-excel-a1
   "Convert zero based column and row number to Excel 'A1' address form"
   [col-num row-num]
-  {:pre [(>= col-num 0) (< col-num 16384)
-         (>= row-num 0) (< row-num 1048576)]}
+  {:pre [(>= col-num 0) (< col-num excel-max-cols)
+         (>= row-num 0) (< row-num excel-max-rows)]}
   (loop [cc (inc col-num)
          acc (conj '() (inc row-num))]
     (if (>= 0 cc)
         (apply str acc)
-      (let [md (mod (- cc 1) 26)]
-        (recur (int (/ (- cc md) 26))
+      (let [md (mod (- cc 1) base-divisor)]
+        (recur (int (/ (- cc md) base-divisor))
                (conj acc (char (+ base-char md))))))))
 
+(defn get-excel-range-a1
+  "Produces a range signature in the form '??:??'"
+  [[sr sc er ec]]
+  (str (get-excel-a1 sr sc) ":" (get-excel-a1 er ec)))
 
+(defn get-data-dimensions
+  "Returns a zero based dimension of the data block and
+  assumes that the first internal collection is representative of all rows
+  data - vector of vectors
+  returns vector of two values: column count and row count"
+  [data]
+  {:pre [(and (vector? data) (vector? (first data)))]}
+  [(dec (count (first data))) (dec (count data))])
+
+(defn pad-rows
+  "Returns a collection with uniform length sub-collections. Uniform Length is
+  determined by max length of sub-collections in coll. Those with length less
+  than max are padded to the right with padval or 0 by default.
+  coll - the input collection of collections
+  padval - pad value (defaults to 0)"
+  [coll & [padval]]
+  (let [basel (into #{} (map count coll))
+        rmax (apply max basel)
+        pad (or padval 0)]
+    (if (<= (count basel) 1)
+      coll
+      (reduce #(conj %1 (vec %2)) (empty coll)
+              (partition
+               rmax
+               (mapcat #(into % (repeat (- rmax (count %)) pad))
+                       coll))))))
+
+;;
 ;; Low level DSL functions ----------------------------------------------------
+;;
+
 
 (def ^:private scrptfile (io/resource "clasew-excel.applescript"))
 
@@ -149,8 +189,10 @@
   {"new_sheet" new_sheet "target" (get sheet-coords target)
    "relative_to" (get sheet-coords relative_to relative_to)})
 
-
+;;
 ;; High level DSL functions ---------------------------------------------------
+;;
+
 
 (defn create-wkbk
   "Creates the script to create new workbook in excel (if not open)
@@ -172,12 +214,9 @@
   (clasew-excel-script bookname no-create open path
                       (clasew-excel-handler chains)))
 
-(defn- get-data-dimensions
-  [data]
-  {:pre [(and (vector? data) (vector? (first data)))]}
-  [(dec (count (first data))) (dec (count data))])
-
+;;
 ;; Chain Sugar ----------------------------------------------------------------
+;;
 
 (defn chain-put-range-data
   "Sets up a chain for putting a data range into workbook's sheet-name. Output
@@ -190,9 +229,8 @@
   (let [sc (or start-col 0)
         sr (or start-row 0)
         [ec er] (get-data-dimensions data)
-        srange (get-excel-a1 sc sr)
-        erange (get-excel-a1 (+ sc ec) (+ sr er))]
-    [:put-range-data sheet-name (str srange ":" erange) data])
+        a1_range (get-excel-range-a1 (list sc sr (+ sc ec) (+ sr er)))]
+    [:put-range-data sheet-name a1_range data])
   )
 
 (defn chain-get-range-data
@@ -235,3 +273,87 @@
   {:pre [(and (not (empty? sheets)) (> (count sheets) 0))]}
   (into [:delete-sheet] sheets))
 
+;;
+;; More Sugar -----------------------------------------------------------------
+;;
+
+(defn formula-wrap
+  "Setup the standard excel '=formula(blah blah blah)' format.
+  form-prefix is one of SUM, AVERAGE, etc.
+  f           function to perform on arg that results in valid formula arguments
+  arg -       whatever arguments are required for the f function to do it's thing"
+  [form-prefix f arg]
+  (str "=" form-prefix "(" (f arg) ")"))
+
+(defn fsum
+  "Thin wrapper for setting up '=SUM(...)'"
+  [f arg]
+  (formula-wrap "SUM" f arg))
+
+(defn favg
+  "Thin wrapper for setting up '=AVERAGE(...)'"
+  [f arg]
+  (formula-wrap "AVERAGE" f arg))
+
+(defn row-ranges
+  "Creates a sequence of full row ranges derived from the data dimensions
+  along with offsets
+  data - a vector of vectors
+  sc - starting column offset (0 based)
+  sr - starting row offset (0 based)"
+  [data & [sc sr]]
+  (let [dc (or sc 0)
+        dr (or sr 0)
+        rwrng (range dr (+ dr (count data)))]
+    (map #(list dc %2 (dec (+ (count %1) dc)) %2) data rwrng)))
+
+(defn column-ranges
+  "Creates a sequence of full column ranges derived from the data dimensions
+  along with offsets
+  data - a vector of vectors
+  sc - starting column offset (0 based)
+  sr - starting row offset (0 based)"
+  [data & [sc sr]]
+  (let [dc (or sc 0)
+        dr (or sr 0)]
+    (map #(list % dr % (dec (+ (count data) dr)))
+         (range dc (+ (count (first data)) dc)))))
+
+(defn sum-by-row
+  "Produces sequence of row sum formulas from input data"
+  [data & [sc sr]]
+  (into [] (map #(fsum get-excel-range-a1 %)
+                (row-ranges data sc sr))))
+
+(defn sum-by-col
+  "Produces sequence of column sum formulas from input data"
+  [data & [sc sr]]
+  (into [] (map #(fsum get-excel-range-a1 %)
+                (column-ranges data sc sr))))
+
+(defn avg-by-row
+  "Produces sequence of row average formulas from input data"
+  [data & [sc sr]]
+  (into [] (map #(favg get-excel-range-a1 %)
+                (row-ranges data sc sr))))
+
+(defn avg-by-col
+  "Produces sequence of row average formulas from input data"
+  [data & [sc sr]]
+  (into [] (map #(favg get-excel-range-a1 %)
+                (column-ranges data sc sr))))
+
+(defn extend-rows
+  "Returns the collection with rows extended to include results of applying one
+  or more functions to the input collection."
+  [coll start-col start-row formfn & formfns]
+  (let [nd (map #(% coll start-col start-row) (conj formfns formfn))]
+    (vec (map #(into %1 %2) coll (apply map vector nd)))
+    ))
+
+(defn extend-columns
+  "Returns the collection with columns extended to include results of applying
+  one or more functions to the input collection."
+  [coll start-col start-row formfn & formfns]
+  (let [nd (map #(% coll start-col start-row) (conj formfns formfn))]
+    (into coll nd)))